Add files via upload
first code that comes from our mexican friends and collects the temperature and dissolved oxygen. We have commented the parts concerning the timing of the data and the saving on SD card.
However it remains a problem as the oxygen values are not correct. This is a reading with 115200 baud rate
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.02
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.02
Temperatura: 17.20
Oxigeno: 0.01
Temperatura: 17.20
Oxigeno: 0.01